Change the code file structure
